test(dsh-notifier): 按功能域拆分 smoke 为 unit/e2e/routes/contract 九个测试文件（#18 方案C）
- smoke.ts 1052 行 -> 薄聚合入口（ESM 声明序保证用例依序执行）
- helpers.ts：fake ctx/req/res、轮询 waitForHistory、agentWithTitle 等共享辅助
- unit-*：config（归一化+免打扰）/text（文案与命令构造）/sanitize（脱敏全套）
- e2e-*：approval/done/interrupt/question-turn（fake ctx + apply，独立临时目录自建自清）
- routes.test.ts：HTTP 路由全套；client-contract.test.ts：两端路由一致性
- 断言总数 265 -> 271（补 /health 405 与配置摘要缺失用例），覆盖不降；
  单测试文件最大 224 行 < 400
